% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{add_loc_info}
\alias{add_loc_info}
\title{Locality Code Lookup
This function reads locality codes and imports locality names
from the locality code dictionary for collection code records.
It ignores records that have no locality code entered.}
\usage{
add_loc_info(df, loc_key)
}
\arguments{
\item{df}{a data frame in standard Cal Academy Collection Code Format}

\item{loc_key}{a data frame file in standard Cal Academy Collection Code Format}
}
\value{
a data frame with updated locality information
}
\description{
Locality Code Lookup
This function reads locality codes and imports locality names
from the locality code dictionary for collection code records.
It ignores records that have no locality code entered.
}
