% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{add_to_filename}
\alias{add_to_filename}
\title{this function streamlines adding clarifying text to the end of a filename}
\usage{
add_to_filename(full.path, text)
}
\arguments{
\item{full.path}{a full path to a file}

\item{text}{the text string you want to add to the file base name}
}
\description{
this function streamlines adding clarifying text to the end of a filename
}
