% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{alpha_filter}
\alias{alpha_filter}
\title{Alpha Character Filter for Strings}
\usage{
alpha_filter(text_vector)
}
\arguments{
\item{text_vector}{A character vector where each element is a string that needs to be filtered to include
only alphabetic characters.}
}
\value{
A character vector of the same length as `text_vector`, where each string is transformed to
        contain only lowercase alphabetic characters.
}
\description{
This function processes a vector of text strings, reducing each to only its lowercase alphabetic characters.
This is useful for simplifying text to its basic elements for more robust regex matching,
eliminating issues with casing, whitespace, and symbols.
}
