% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{bi_tri}
\alias{bi_tri}
\title{This function is used to generate a binomial and trinomial column based on
Pre-existing genus and species columns}
\usage{
bi_tri(df)
}
\arguments{
\item{df}{a data frame of occurrence records in Symbiota format}
}
\value{
a data frame with binomial and trinomial columns
}
\description{
This function is used to generate a binomial and trinomial column based on
Pre-existing genus and species columns
}
