% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{bracketr}
\alias{bracketr}
\title{function that adds brackets appropriately around lists that are nested in each element in a column}
\usage{
bracketr(column)
}
\arguments{
\item{column}{a data frame column}
}
\value{
a modified data frame
}
\description{
function that adds brackets appropriately around lists that are nested in each element in a column
}
