% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{cargo_data_ns_formatr}
\alias{cargo_data_ns_formatr}
\title{function that formats a data frame for importation using the Data Transfer extension
so that content can be uploaded to MediaWiki}
\usage{
cargo_data_ns_formatr(df, title = "title", free_text = "free_text", prefix)
}
\arguments{
\item{df}{a frame to be formatted}

\item{title}{the "Title" column for Cargo}

\item{free_text}{the "Free Text" column for Cargo}

\item{prefix}{a string with the prefix for each field which corresponds to the table name in Cargo}
}
\description{
function that formats a data frame for importation using the Data Transfer extension
so that content can be uploaded to MediaWiki
}
