% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{check_empty}
\alias{check_empty}
\title{Checks for missing data in key collection fields
This function takes a standard critter club data frame
and checks for blanks in missing fields}
\usage{
check_empty(df)
}
\arguments{
\item{df}{standard Cal Academy Insect Dataframe for writing insect labels}
}
\description{
Checks for missing data in key collection fields
This function takes a standard critter club data frame
and checks for blanks in missing fields
}
