% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{clean_ent_label}
\alias{clean_ent_label}
\title{Add proper abbreviations, spacing, and punction for insect labels
This function takes a label inbox object and writes
appropriate label format in text file to serve as input
for latex'}
\usage{
clean_ent_label(label_inbox)
}
\arguments{
\item{label_inbox}{standard Cal Academy Insect Dataframe for writing insect labels}
}
\value{
an edited dataframe in preparation for inputing to LaTeX
}
\description{
Add proper abbreviations, spacing, and punction for insect labels
This function takes a label inbox object and writes
appropriate label format in text file to serve as input
for latex'
}
