% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{clean_ento_fields}
\alias{clean_ento_fields}
\title{streamline word spellings for CASENT style insect database}
\usage{
clean_ento_fields(df)
}
\arguments{
\item{df}{a database in CASENT format}
}
\value{
an updated dataframe in CASENT format
}
\description{
streamline word spellings for CASENT style insect database
}
