% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{common_name_caps}
\alias{common_name_caps}
\title{this is a custom function to fix common name capitalization nuances
i.e. title casing but not capitalizing after hyphens}
\usage{
common_name_caps(vec)
}
\arguments{
\item{vec}{a vector of common names of organisms}
}
\value{
a vector with results
}
\description{
this is a custom function to fix common name capitalization nuances
i.e. title casing but not capitalizing after hyphens
}
