% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{common_namer}
\alias{common_namer}
\title{Wrapper function to query rgbif for common names and append them to biological
occurrence records. Functions takes a data frame of records and attaches
common names using rgif in English and Spanish. Note that common names have to
be queried by scientific name key numbers and not from strings of names.
For convenience, the usageKey field returned by gbif is the numerical key
referring to the lowest ranked taxonomic name for convenience}
\usage{
common_namer(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
Wrapper function to query rgbif for common names and append them to biological
occurrence records. Functions takes a data frame of records and attaches
common names using rgif in English and Spanish. Note that common names have to
be queried by scientific name key numbers and not from strings of names.
For convenience, the usageKey field returned by gbif is the numerical key
referring to the lowest ranked taxonomic name for convenience
}
