% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{deparse_codes}
\alias{deparse_codes}
\title{removes the base and subcode columns that the function parse codes adds}
\usage{
deparse_codes(df)
}
\arguments{
\item{df}{Either a collection code or speciment code database in CASENT format}
}
\value{
an updated database
}
\description{
removes the base and subcode columns that the function parse codes adds
}
