% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{df_to_mw_structure}
\alias{df_to_mw_structure}
\title{Function that takes a data frame and creates one field containing data from all other fields
for each row that are in MediaWiki data storage format. This format is used to make calls
to the MediaWiki Cargo extension to produce a Cargo database without needed separate template
calls stored on individual wiki pages}
\usage{
df_to_mw_structure(df, cargo_table, target_column, col_changes = NA)
}
\arguments{
\item{df}{a data frame}

\item{cargo_table}{the name of the Cargo table to make a call to}

\item{target_column}{the name of the R field to store the data}

\item{col_changes}{a comma separated list representing strings to be treated as abbreviations by snakecase function}
}
\value{
a modified data frame
}
\description{
Function that takes a data frame and creates one field containing data from all other fields
for each row that are in MediaWiki data storage format. This format is used to make calls
to the MediaWiki Cargo extension to produce a Cargo database without needed separate template
calls stored on individual wiki pages
}
