% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{ento_dates}
\alias{ento_dates}
\title{Convert to standard insect dates for labels}
\usage{
ento_dates(date_vector_start, date_vector_end)
}
\arguments{
\item{date_vector_start}{a vector of start dates}

\item{date_vector_end}{a vector of end dates}
}
\value{
a dataframe column of date or date range that collection took place
}
\description{
Convert to standard insect dates for labels
}
