% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{extract_coords}
\alias{extract_coords}
\title{function which geoencodes records based on date and time stamps with gpx files}
\usage{
extract_coords(records, timeCoords, timezone)
}
\arguments{
\item{records}{a data frame of occurence records that contains lat and lon columns}

\item{timeCoords}{a csv of coordinates with time stamps usually from gps unit}

\item{timezone}{timezone in which to interpret the biological records for matching to timeCoords (which are in UTC)}
}
\value{
a dataframe with filled in lat lon values
}
\description{
function which geoencodes records based on date and time stamps with gpx files
}
