% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{fill_taxa}
\alias{fill_taxa}
\title{function that fills in missing taxon information—generates a dataframe
with same format as original dataframe needs to either have a 'knowntaxalevel'
and 'taxonidinput' column, or dataframe needs to have the following columns:
'class', 'order', 'family', 'subfamily', 'genus', 'species'. function outputs
a dataframe with same row and column order as original dataframe.}
\usage{
fill_taxa(df, key)
}
\arguments{
\item{df}{a database in CASENT format}

\item{key}{# key containing taxonomic information}
}
\description{
function that fills in missing taxon information—generates a dataframe
with same format as original dataframe needs to either have a 'knowntaxalevel'
and 'taxonidinput' column, or dataframe needs to have the following columns:
'class', 'order', 'family', 'subfamily', 'genus', 'species'. function outputs
a dataframe with same row and column order as original dataframe.
}
