% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{filter_time_diff}
\alias{filter_time_diff}
\title{this function filters out records based on time difference between observations
it assumes the data is sorted first by a unique id vector and then by ascending date/time}
\usage{
filter_time_diff(obs, ind = 1)
}
\arguments{
\item{obs}{observation database in standard critter club format}

\item{ind}{need to document}
}
\description{
this function filters out records based on time difference between observations
it assumes the data is sorted first by a unique id vector and then by ascending date/time
}
