% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{find_bad_G}
\alias{find_bad_G}
\title{Checks that all genera listed in a set of observations occurs in the specified key or checklist}
\usage{
find_bad_G(obs, key)
}
\arguments{
\item{obs}{observation databse in standard critter club format}

\item{key}{key or checklist in standard critter club format}
}
\description{
Checks that all genera listed in a set of observations occurs in the specified key or checklist
}
