% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{find_elevation}
\alias{find_elevation}
\title{Populate elevation using the geonames server based on
lat and lon columns in a dataframe
# I obtained my username (which is 'roverso') from here to validate the elevation lookup request
http://www.geonames.org/export/web-services.html}
\usage{
find_elevation(df)
}
\arguments{
\item{df}{a database in SYMBIOTA format}
}
\value{
a dataframe with rounded lat lon values
}
\description{
Populate elevation using the geonames server based on
lat and lon columns in a dataframe
# I obtained my username (which is 'roverso') from here to validate the elevation lookup request
http://www.geonames.org/export/web-services.html
}
