% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{find_max_sc}
\alias{find_max_sc}
\title{Last specimen code finder
function which takes a collection code and db_sc and returns
the last specimen code (highest number) associated with that collection.
This function is useful for making sure that specimen codes are not duplicated
when generating new ones.}
\usage{
find_max_sc(cc.code, db_sc)
}
\arguments{
\item{cc.code}{a numeric value corresponding to a collection code}

\item{db_sc}{a data frame file in standard Cal Academy specimen code database format}
}
\value{
a numeric value corresponding to the last assigned specimen code for a given collection code
}
\description{
Last specimen code finder
function which takes a collection code and db_sc and returns
the last specimen code (highest number) associated with that collection.
This function is useful for making sure that specimen codes are not duplicated
when generating new ones.
}
