% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{first_finder}
\alias{first_finder}
\title{Function that finds the first non-blank entry in a vector for use with dplyr::rowwise()
The suppressWarnings is for warnings when their are no elements in the vector}
\usage{
first_finder(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the value of the first non-blank entry in the vector
}
\description{
Function that finds the first non-blank entry in a vector for use with dplyr::rowwise()
The suppressWarnings is for warnings when their are no elements in the vector
}
