% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{flatten_rgbif}
\alias{flatten_rgbif}
\title{Flattens the nested structure of a rgbif backbone_name query
and streamlines other inconsistencies in data structure}
\usage{
flatten_rgbif(df, result_column)
}
\arguments{
\item{df}{a df with a column that contains nested query response}

\item{result_column}{name of result column with response from rgbif::backbone_name()}
}
\value{
a data frame
}
\description{
Flattens the nested structure of a rgbif backbone_name query
and streamlines other inconsistencies in data structure
}
