% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{generic_wiki_formatter}
\alias{generic_wiki_formatter}
\title{This function in deprecated. It takes a processed data frame from airtable_reader function and further formats
it for uploading as structured data to MediaWiki}
\usage{
generic_wiki_formatter(df, cargo_template_name, cargo = TRUE, col_changes = NA)
}
\arguments{
\item{df}{a data frame with at least one column called "name" that represent the name of the wiki page}

\item{cargo_template_name}{a string representing the name of the cargo_template's name associated with the upload}

\item{cargo}{either TRUE or FALSE depending on whether formatting a cargo template call is desired}

\item{col_changes}{a comma separated list representing strings to be treated as abbreviations by snake case function}
}
\description{
This function in deprecated. It takes a processed data frame from airtable_reader function and further formats
it for uploading as structured data to MediaWiki
}
