% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{generic_wiki_formatter2}
\alias{generic_wiki_formatter2}
\title{This updated version of the wiki formatter has additional features. It takes a processed data frame from the airtable_reader
function and further formats it for uploading as structured data to MediaWiki}
\usage{
generic_wiki_formatter2(
  df,
  wiki_data_structure,
  single_data_page = NULL,
  data_table_prefix = NULL,
  tags_and_templates = NULL,
  cargo_template_name,
  cargo = TRUE,
  data_page_bread_crumb = FALSE,
  forward_facing = FALSE,
  col_changes = NA
)
}
\arguments{
\item{df}{a data frame with at least one column called "name" that represent the name of the wiki page}

\item{wiki_data_structure}{a string representing the desired data structure on the wiki. Options are "one-to-one" or "collapsed"}

\item{single_data_page}{= NULL, the name of the data page if all records are to go on a single data page}

\item{tags_and_templates}{A data frame containing the columns that contain the tags and templates for the associated data}

\item{cargo_template_name}{A string representing the name of the cargo_template's name associated with the upload}

\item{cargo}{= TRUE,  Either TRUE or FALSE depending on whether formatting a cargo template call is desired}

\item{forward_facing}{= FALSE, Either TRUE or FALSE depending on whether the data will end up on a forward-facing page
or whether it will be transcluded by another page. In the latter case it needs to be wrapped in the appropriate tags. For now
forward_facing set to TRUE only works with a wiki_data_structure of one-to-one.}

\item{col_changes}{a comma separated list representing strings to be treated as abbreviations by snake case function}
}
\description{
This updated version of the wiki formatter has additional features. It takes a processed data frame from the airtable_reader
function and further formats it for uploading as structured data to MediaWiki
}
