% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{grab_section_headers}
\alias{grab_section_headers}
\title{This function is used with text parsing to grab a header that is up above
an entry and associate it with all entries below it}
\usage{
grab_section_headers(base_text_loc, header_loc, input_vec)
}
\arguments{
\item{base_text_loc}{the locations (row numbers) of the entries}

\item{header_loc}{the locations (row numbers) of the headers}

\item{input_vec}{the input text vector that contains both of the above}
}
\value{
a data frame
}
\description{
This function is used with text parsing to grab a header that is up above
an entry and associate it with all entries below it
}
