% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{graph_at_taxon}
\alias{graph_at_taxon}
\title{This function plots family observation level results by type of animal. The
aes_string function was an elegant way to have user input become col_names
within a ggplot wrapper}
\usage{
graph_at_taxon(checklist, type, taxonRank, counts, dir)
}
\value{
a vector with results
}
\description{
This function plots family observation level results by type of animal. The
aes_string function was an elegant way to have user input become col_names
within a ggplot wrapper
}
