% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{gsub2}
\alias{gsub2}
\title{modified gsub function that allows substitution of lists of key value pairs}
\usage{
gsub2(pattern, replacement, x, ...)
}
\description{
modified gsub function that allows substitution of lists of key value pairs
}
