% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{higher_rank_remover}
\alias{higher_rank_remover}
\title{Function that removes higher linnean rank names from the scientific_name column
if they are already represented by lower ranked names, for example, removing
'Lepidoptera' if 'Danaus plexippus' already exists in the same column.}
\usage{
higher_rank_remover(df, rankList)
}
\arguments{
\item{df}{a dataframe}

\item{rankList}{a list of linnean rank names}
}
\value{
a data frame
}
\description{
Function that removes higher linnean rank names from the scientific_name column
if they are already represented by lower ranked names, for example, removing
'Lepidoptera' if 'Danaus plexippus' already exists in the same column.
}
