% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{include_only_prefix_maker}
\alias{include_only_prefix_maker}
\title{This function produces the 'prefix' file necessary to preface MediaWiki subsequently
uploaded content with a <noinclude> tag. This is a silly 'hack' file that needs
to be uploaded before the template calls on the data pages so that the calls are
sandwiched in an <includeonly> tag pair. It literally just uploads the text "<includeonly>" as a precursor before
any other data is appended to the page#'}
\usage{
include_only_prefix_maker(df, title = "Title", output_path)
}
\arguments{
\item{df}{a data frame}

\item{title}{the column that contains the names of wiki pages to be processed}

\item{output_path}{the path to save the prefix file which will later be uploaded to a MediaWiki instance}
}
\description{
This function produces the 'prefix' file necessary to preface MediaWiki subsequently
uploaded content with a <noinclude> tag. This is a silly 'hack' file that needs
to be uploaded before the template calls on the data pages so that the calls are
sandwiched in an <includeonly> tag pair. It literally just uploads the text "<includeonly>" as a precursor before
any other data is appended to the page#'
}
