% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{linnean_releveler}
\alias{linnean_releveler}
\title{Function to sort biological occurrence records by linnean ranks}
\usage{
linnean_releveler(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
Function to sort biological occurrence records by linnean ranks
}
