% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{lookup_common_name}
\alias{lookup_common_name}
\title{Custom function to do table join to fill in data from common name lists
probably don't need this function anymore
i.e. title casing but not capitalizing after hyphens}
\usage{
lookup_common_name(df, checklist)
}
\arguments{
\item{df}{a dataframe}

\item{checklist}{# a checklist}
}
\value{
a data frame with results
}
\description{
Custom function to do table join to fill in data from common name lists
probably don't need this function anymore
i.e. title casing but not capitalizing after hyphens
}
