% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{lua_df_converter}
\alias{lua_df_converter}
\title{Function which converts a data frame/tibble to WikiMedia Lua Module format}
\usage{
lua_df_converter(df, id_column)
}
\arguments{
\item{df}{a data frame or tibble}

\item{id_column}{a string containing the name of the unique (leftmost) id column}
}
\value{
a vector with results
}
\description{
Function which converts a data frame/tibble to WikiMedia Lua Module format
}
