% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{lua_dict_builder}
\alias{lua_dict_builder}
\title{function that rearranges a dataframe to a lookup table style format where every
unique value across focal columns is represented in the leftmost column of the dataframe}
\usage{
lua_dict_builder(df, col_vec)
}
\arguments{
\item{df}{a data frame or tibble}

\item{col_vec}{a list of strings corresponding to the focal columns to be rearranged}
}
\value{
a vector with results
}
\description{
function that rearranges a dataframe to a lookup table style format where every
unique value across focal columns is represented in the leftmost column of the dataframe
}
