% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{make_latex_source}
\alias{make_latex_source}
\title{Creates latex source data files for each unique class in a data frame.}
\usage{
make_latex_source(df, rank_vec, fp_out)
}
\arguments{
\item{df}{a data frame}

\item{fp_out}{a file path for the latex output to export to}

\item{rank}{a string list representing a linnean rank system}
}
\description{
Creates latex source data files for each unique class in a data frame.
}
