% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{merge_with_order}
\alias{merge_with_order}
\title{function that keeps original order after merging data frames}
\usage{
merge_with_order(x, y, ..., sort = T, keep_order)
}
\description{
function that keeps original order after merging data frames
}
