% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{mi_to_km}
\alias{mi_to_km}
\title{function that finds miles in a string of plain English
and converts to km in a string}
\usage{
mi_to_km(vec)
}
\arguments{
\item{vec}{a vector of values in miles}
}
\value{
a vector of values in kilometers
}
\description{
function that finds miles in a string of plain English
and converts to km in a string
}
