% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{munge_gpx}
\alias{munge_gpx}
\title{Reads in gpx files recursively from a designated path and assembles and formats one
exhaustive dataframe with just location and time stamps sorted by time}
\usage{
munge_gpx(path)
}
\arguments{
\item{path}{a path to parent folder where gpx files occur}
}
\value{
a data frame of coords and time stamps for all gpx files specified
}
\description{
Reads in gpx files recursively from a designated path and assembles and formats one
exhaustive dataframe with just location and time stamps sorted by time
}
