% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{na_skipper}
\alias{na_skipper}
\title{Given a dataframe and vector of column names ranked in the order of highest to lowest priority,
this function will mutate the column 'first' which will contain the first occuring non-NA value among the
columns specified. Note that the syms() function is required to convert strings to variables}
\usage{
na_skipper(df, rank_vec)
}
\arguments{
\item{df}{a data frame}

\item{rank_vec}{a data frame}
}
\value{
a data frame
}
\description{
Given a dataframe and vector of column names ranked in the order of highest to lowest priority,
this function will mutate the column 'first' which will contain the first occuring non-NA value among the
columns specified. Note that the syms() function is required to convert strings to variables
}
