% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{parse_codes}
\alias{parse_codes}
\title{Splits up an insect collection dataframe's (collection or specimen dataframe) code column
into constituent parts: base code number and subcode (if present).
used to match specimen codes with their parent collection codes efficiently}
\usage{
parse_codes(df)
}
\arguments{
\item{df}{Either a collection code or speciment code database in CASENT format}
}
\value{
an updated database
}
\description{
Splits up an insect collection dataframe's (collection or specimen dataframe) code column
into constituent parts: base code number and subcode (if present).
used to match specimen codes with their parent collection codes efficiently
}
