% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{pasteSansNA}
\alias{pasteSansNA}
\title{Paste function that doesn't carry over NA values}
\usage{
pasteSansNA(..., sep = ", ")
}
\value{
a vector with results
}
\description{
Paste function that doesn't carry over NA values
}
