% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{pre_query_patchr}
\alias{pre_query_patchr}
\title{function that makes manual edits based on a csv key before records are queried
at gbif}
\usage{
pre_query_patchr(df, band_aid)
}
\arguments{
\item{df}{a data frame}

\item{band_aid}{a data frame}
}
\value{
a data frame
}
\description{
function that makes manual edits based on a csv key before records are queried
at gbif
}
