% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{process_source}
\alias{process_source}
\title{Control Execution of a Source Script Based on Data Modifications}
\usage{
process_source(raw_data_paths, source_path)
}
\arguments{
\item{raw_data_paths}{A vector of file paths pointing to the raw data files. These paths are
checked to see if any modifications have been made that necessitate re-running the
processing script.}

\item{source_path}{A string specifying the path to the source file that may be executed if
the data has changed since the last run.}
}
\description{
This function checks if the raw data specified by `raw_data_paths` has been modified since
the last successful run (compared to a timestamp). If the data is newer, it executes a script
located at `source_path`, updates the timestamp, and informs the user of completion. If the data
has not changed, it notifies the user that processing will be skipped.
}
