% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{rds_path_maker}
\alias{rds_path_maker}
\title{Generate RDS File Path from Source Path}
\usage{
rds_path_maker(source_path)
}
\arguments{
\item{source_path}{A character string specifying the path to the source file.}
}
\description{
Constructs a file path for an RDS file based on the provided source file path.
The resulting path includes the base file name of the source path (without extension)
and appends '.rds' extension, storing the file in an 'rds' directory relative to the project root.
}
