% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{rds_timestamp_path_maker}
\alias{rds_timestamp_path_maker}
\title{Generate Path for RDS Timestamp File from Source Path}
\usage{
rds_timestamp_path_maker(source_path)
}
\arguments{
\item{source_path}{A character string specifying the path to the source file.}
}
\value{
A character string with the full path to the corresponding RDS timestamp file in the 'rds' directory.
}
\description{
Constructs a file path for an RDS timestamp file based on the provided source file path.
The function appends '_edit_time.rds' to the base file name of the source path,
indicating that the RDS file is used to store modification timestamps.
The file is stored in an 'rds' directory relative to the project root.
}
