% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_cc_db}
\alias{read_cc_db}
\title{Read Collection Code tsv
This function reads in a tsv in the standard
Cal Academy Collection data format with correct
data types assigned}
\usage{
read_cc_db(collectionDBFilePath)
}
\arguments{
\item{collectionDBFilePath}{a tsv file in standard Cal Academy Collection Code Format}
}
\value{
a data frame file in standard Cal Academy Collection Code Format
}
\description{
Read Collection Code tsv
This function reads in a tsv in the standard
Cal Academy Collection data format with correct
data types assigned
}
