% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_csv_critter}
\alias{read_csv_critter}
\title{Wrapper for readr::read_csv that forces all cols to character with a few exceptions
pertinent to critter club which also avoids the erroneous auto-guessing
errors stemming from fields which have too many NA's in their records
requires date fields to be in Symbiota format}
\usage{
read_csv_critter(csv_path)
}
\arguments{
\item{csv_path}{a path to a csv file#'}
}
\description{
Wrapper for readr::read_csv that forces all cols to character with a few exceptions
pertinent to critter club which also avoids the erroneous auto-guessing
errors stemming from fields which have too many NA's in their records
requires date fields to be in Symbiota format
}
