% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_dates_safely}
\alias{read_dates_safely}
\title{function that safely reads in CASENT insect database and handles dates with lubridate
reads in all other columns as character/text}
\usage{
read_dates_safely(path, sheet.num)
}
\arguments{
\item{path}{path to CASENT insect database in Excel format}

\item{sheet.num}{sheet number that database is located on}
}
\value{
an XLS database with dates that are not messed up
}
\description{
function that safely reads in CASENT insect database and handles dates with lubridate
reads in all other columns as character/text
}
