% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_label_inbox}
\alias{read_label_inbox}
\title{Read Label Inbox tsv
This function reads in a tsv in the standard
Cal Academy Specimen data format with correct
data types assigned}
\usage{
read_label_inbox(specimenDBFilePath)
}
\arguments{
\item{specimenDBFilePath}{a tsv file in standard Cal Academy Specimen Code Format}
}
\value{
a data frame file in standard Cal Academy Specimen Code Format
}
\description{
Read Label Inbox tsv
This function reads in a tsv in the standard
Cal Academy Specimen data format with correct
data types assigned
}
