% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_sc_db}
\alias{read_sc_db}
\title{Read Specimen Code tsv
This function reads in a tsv in the standard
Cal Academy Specimen data format with correct
data types assigned}
\usage{
read_sc_db(specimenDBFilePath)
}
\arguments{
\item{specimenDBFilePath}{a tsv file in standard Cal Academy Specimen Code Format}
}
\value{
a data frame file in standard Cal Academy Specimen Code Format
}
\description{
Read Specimen Code tsv
This function reads in a tsv in the standard
Cal Academy Specimen data format with correct
data types assigned
}
