% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_xlsx_char}
\alias{read_xlsx_char}
\title{wrapper for readxl::read_xlsx() that forces all cols to character
which also avoids the very stupid erroneous autoguessing errors stemming from
fields which have too many NA's in their records}
\usage{
read_xlsx_char(xlsx_path)
}
\arguments{
\item{xlsx_path}{a path to an xlsx file}
}
\description{
wrapper for readxl::read_xlsx() that forces all cols to character
which also avoids the very stupid erroneous autoguessing errors stemming from
fields which have too many NA's in their records
}
