% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{read_xlsx_critter}
\alias{read_xlsx_critter}
\title{Wrapper for readr::read_xlsx that forces all cols to character with a few exceptions
pertinent to critter club and formats columns for critter club records}
\usage{
read_xlsx_critter(xlsx_path, sheet, na = "")
}
\arguments{
\item{xlsx_path}{a path to a csv file#'}

\item{sheet}{an integer representing the sheet number to read}
}
\description{
Wrapper for readr::read_xlsx that forces all cols to character with a few exceptions
pertinent to critter club and formats columns for critter club records
}
