% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{rev_geocode_insect}
\alias{rev_geocode_insect}
\title{reverse geocode function. This function takes a standard Cal Academy
database and uses the lat lon columns to add additional reverse
geocoded information if it is not already filled in.}
\usage{
rev_geocode_insect(df, api)
}
\arguments{
\item{df}{a standard Cal Academy Insect Database format for insect labels}

\item{api}{a Google api to access the reverse geocode data}
}
\value{
returns a database with reverse geocoded fields
}
\description{
reverse geocode function. This function takes a standard Cal Academy
database and uses the lat lon columns to add additional reverse
geocoded information if it is not already filled in.
}
