% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{round_coords_cc}
\alias{round_coords_cc}
\title{Rounds lat lon values based on sensible precision and manage formats
based on accuracy information in an coordinate uncertainty field
in meters}
\usage{
round_coords_cc(df)
}
\arguments{
\item{df}{a database in SYMBIOTA format}
}
\value{
a dataframe with rounded lat lon values
}
\description{
Rounds lat lon values based on sensible precision and manage formats
based on accuracy information in an coordinate uncertainty field
in meters
}
