% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{round_coords_insect}
\alias{round_coords_insect}
\title{rounds lat lon values based on sensible precision and manage formats
based on accuracy information field records}
\usage{
round_coords_insect(df)
}
\arguments{
\item{df}{a database in CASENT format}
}
\value{
a dataframe with rounded lat lon values
}
\description{
rounds lat lon values based on sensible precision and manage formats
based on accuracy information field records
}
