% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{safe_ifelse}
\alias{safe_ifelse}
\title{ifelse function that doesn't obliterate classes (i.e. date formats, etc...)}
\usage{
safe_ifelse(cond, yes, no)
}
\description{
ifelse function that doesn't obliterate classes (i.e. date formats, etc...)
}
