% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{sci_name_parser}
\alias{sci_name_parser}
\title{function that takes biological records that only have a scientific_name column
and parses it and adds Linnaean columns: kingdom, phylum, class, order,
family, genus...}
\usage{
sci_name_parser(df)
}
\arguments{
\item{df}{a data frame of occurrence records in Symbiota format}
}
\value{
a data frame with expanded Linnaean columns
}
\description{
function that takes biological records that only have a scientific_name column
and parses it and adds Linnaean columns: kingdom, phylum, class, order,
family, genus...
}
