% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{split_cc_code}
\alias{split_cc_code}
\title{this function receives collection code records in standard format
and splits them into specimen code records based on user input and then
adds them to the specimen code master database}
\usage{
split_cc_code(cc.records, db_sc)
}
\arguments{
\item{cc.records}{a dataframe of collection codes}

\item{db_sc}{a specimen database}
}
\value{
split specimen records
}
\description{
this function receives collection code records in standard format
and splits them into specimen code records based on user input and then
adds them to the specimen code master database
}
