% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{stack_writer}
\alias{stack_writer}
\title{Write Unique Non-empty Values from Dataframe to File}
\usage{
stack_writer(df, file_path)
}
\arguments{
\item{df}{A dataframe with any number of columns and rows. The function will
consider all columns and convert the dataframe to a long format.}

\item{file_path}{A character string specifying the full path and filename where
the unique, non-empty values will be saved.}
}
\description{
Converts a dataframe into a long format, filters out all duplicate and empty or NA values,
and writes the remaining unique values to a specified file. This function is useful for
creating lists or stacks of unique entries from tabular data.
}
