% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{symb_tidy_rosetta}
\alias{symb_tidy_rosetta}
\title{Renaming keys for "Rosetta stone" that converts data frame names from
Symbiota/Darwin Core format to tidy format}
\usage{
symb_tidy_rosetta(df, direction = "forward")
}
\arguments{
\item{df}{a data frame}

\item{direction}{a string containing either "forward", the default, for Symbiota to Tidy conversion or "reverse" for the oppotite}
}
\description{
Renaming keys for "Rosetta stone" that converts data frame names from
Symbiota/Darwin Core format to tidy format
}
