% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{tag_by_poly}
\alias{tag_by_poly}
\title{Function which filters coordinates in a data frame based on whether they
fall within a polygon}
\usage{
tag_by_poly(
  df,
  poly,
  lat = "lat",
  lon = "lon",
  crs,
  results_col = "results",
  tag
)
}
\arguments{
\item{df}{a data frame of occurrence records that contains lat and lon columns}

\item{poly}{a shape files with polygons}

\item{lat}{decimal latitude field from the data from}

\item{lon}{decimal longitude field from the data from}

\item{crs}{a projection system}

\item{results_col}{a character string representing the column where you like the label to be stored that indicates selected points}

\item{tag}{a character string with which to label selected points}
}
\value{
a data frame with records filtered based on falling inside the focal polygon
}
\description{
Function which filters coordinates in a data frame based on whether they
fall within a polygon
}
