% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{taxo_updatr}
\alias{taxo_updatr}
\title{This function matches a Symbiota formatted data frame of biological occurrences to
a local gbif dictionary created by gbif_dictionary_builder.R and updates taxonomic
names accordingly while warning the user about names that cannot be matched}
\usage{
taxo_updatr(df, dict)
}
\arguments{
\item{df}{a data frame}

\item{dict}{a data frame with a look-up dictionary}
}
\value{
a data frame
}
\description{
This function matches a Symbiota formatted data frame of biological occurrences to
a local gbif dictionary created by gbif_dictionary_builder.R and updates taxonomic
names accordingly while warning the user about names that cannot be matched
}
