% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{taxon.summ.by.type}
\alias{taxon.summ.by.type}
\title{This function takes a filled checklist, a type of animal, and a taxon rank
 and makes a standard critter club checklist report}
\usage{
taxon.summ.by.type(filled.cl, type, taxon.rank)
}
\arguments{
\item{filled.cl}{a filled checklist}

\item{type}{a type of organism}

\item{taxon.rank}{a taxonomic rank}
}
\value{
a vector with results
}
\description{
This function takes a filled checklist, a type of animal, and a taxon rank
 and makes a standard critter club checklist report
}
