% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{update_sc_from_cc}
\alias{update_sc_from_cc}
\title{pseudo-relational database update for CASENT collection and specimen code databases
PSEUDO RELATIONAL DATABASE: OVERWRITE DISCREPANCIES IN SPECIMEN CODE DATABASE db_cc -> db_sc}
\usage{
update_sc_from_cc(db_sc, db_cc)
}
\arguments{
\item{db_sc}{Specimen code database in CASENT format}

\item{db_cc}{Collection code database in CASENT format}
}
\value{
an updated specimen database
}
\description{
pseudo-relational database update for CASENT collection and specimen code databases
PSEUDO RELATIONAL DATABASE: OVERWRITE DISCREPANCIES IN SPECIMEN CODE DATABASE db_cc -> db_sc
}
