% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{updateTimestamp}
\alias{updateTimestamp}
\title{Update the RDS Timestamp After Data Processing}
\usage{
updateTimestamp(raw_data_path, source_path)
}
\arguments{
\item{raw_data_path}{A character string specifying the path to the raw data file.
This is used to retrieve the last modification time for logging or comparison
purposes.}

\item{source_path}{A character string that specifies the path to the source file. This path
is used to construct the path to the RDS timestamp file where the current
system time will be saved.}
}
\description{
This function updates the modification timestamp stored in an RDS file to reflect
the current system time. This operation typically follows the rerunning of source code
that processes raw data, especially when the raw data is determined to be newer than
the timestamp recorded in the RDS file. The function assumes that `rds_path_maker`
and `rds_timestamp_path_maker` are available to construct paths to the RDS files.
}
