% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{wiki_page_protector}
\alias{wiki_page_protector}
\title{Protect Wiki Pages from Overwrite
This function filters out wiki pages that already exist to prevent overwriting.
It first retrieves a list of all existing pages from the specified MediaWiki site.
Each page name is processed to remove non-alphabetic characters for a cleaner comparison.
Finally, it determines which pages from the provided data frame do not exist on the wiki
and returns these as a new dataframe.}
\usage{
wiki_page_protector(pages_df, site_py)
}
\arguments{
\item{pages_df}{A dataframe containing at least one column `page_name` which includes
the names of pages intended for upload or verification.}

\item{site_py}{A Python object representing the MediaWiki site, interfaced through reticulate.}
}
\value{
A dataframe containing the subset of `pages_df` where page names do not exist
        on the wiki site, excluding the intermediate regex-processed names.
}
\description{
Protect Wiki Pages from Overwrite
This function filters out wiki pages that already exist to prevent overwriting.
It first retrieves a list of all existing pages from the specified MediaWiki site.
Each page name is processed to remove non-alphabetic characters for a cleaner comparison.
Finally, it determines which pages from the provided data frame do not exist on the wiki
and returns these as a new dataframe.
}
