% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{write.critter.kml}
\alias{write.critter.kml}
\title{This is a wrapper of the writeOGR function that takes a critter club
database format and turns it to a spatial points data frame
and then converts it to a kml for viewing in Google Earth
There is a bug in the KML driver for writeOGR that unfortunately throws a cryptic warning
but the code seems to still work}
\usage{
write.critter.kml(df, output.path)
}
\arguments{
\item{df}{a data frame in standard Cal Academy Collection Code Format}

\item{output.path}{the full path of where the file should be saved}
}
\description{
This is a wrapper of the writeOGR function that takes a critter club
database format and turns it to a spatial points data frame
and then converts it to a kml for viewing in Google Earth
There is a bug in the KML driver for writeOGR that unfortunately throws a cryptic warning
but the code seems to still work
}
