% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{write_csv_critter}
\alias{write_csv_critter}
\title{Wrapper for readr::write_csv that formats tidy formatted data frames
back to Symbiota format before writing them off}
\usage{
write_csv_critter(df, csv_path)
}
\arguments{
\item{df}{a data frame to write to csv}

\item{csv_path}{a path to save to}
}
\description{
Wrapper for readr::write_csv that formats tidy formatted data frames
back to Symbiota format before writing them off
}
