% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectable.R
\name{write_latex_labels}
\alias{write_latex_labels}
\title{Write insect labels
This function takes a label inbox tsv file and writes
appropriate label format in text file to serve as input
for latex to build insect labels}
\usage{
write_latex_labels(label_inbox)
}
\arguments{
\item{label_inbox}{standard Cal Academy Insect Dataframe for writing insect labels}
}
\value{
A text output readable by latex
}
\description{
Write insect labels
This function takes a label inbox tsv file and writes
appropriate label format in text file to serve as input
for latex to build insect labels
}
